+++ CHANGELOG.md
@@ -41,6 +41,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add testing utils for plugins
 - Add `cache` field in Component
 - Add predict_id in Listener
+- Add serve in Model
 
 #### Bug Fixes